+++ R/dydx.R
@@ -15,7 +15,7 @@
 #' \code{dydx} is an S3 generic with classes implemented for specific variable types. S3 method dispatch, somewhat atypically, is based upon the class of \code{data[[variable]]}.
 #' 
 #' For numeric (and integer) variables, the method calculates an instantaneous marginal effect using a simple \dQuote{central difference} numerical differentiation:
-#' \deqn{\frac{f(x + \frac{1}{2}h) - f(x - \frac{1}{2}h}){dh}}{(f(x + 0.5h) - f(x - 0.5h))/(2h)}, where (\eqn{h = \max(|x|, 1) \sqrt{\epsilon}}{h = max(|x|, 1)sqrt(epsilon)} and the value of \eqn{\epsilon}{epsilon} is given by argument \code{eps}. This procedure is subject to change in the future.
+#' \deqn{\frac{f(x + \frac{1}{2}h) - f(x - \frac{1}{2}h)}{dh}}{(f(x + 0.5h) - f(x - 0.5h))/(2h)}, where (\eqn{h = \max(|x|, 1) \sqrt{\epsilon}}{h = max(|x|, 1)sqrt(epsilon)} and the value of \eqn{\epsilon}{epsilon} is given by argument \code{eps}. This procedure is subject to change in the future.
 #' 
 #' For factor variables (or character variables, which are implicitly coerced to factors by modelling functions), discrete first-differences in predicted outcomes are reported instead (i.e., change in predicted outcome when factor is set to a given level minus the predicted outcome when the factor is set to its baseline level). These are sometimes called \dQuote{partial effects}. If you want to use numerical differentiation for factor variables (which you probably do not want to do), enter them into the original modelling function as numeric values rather than factors.
 #' 
@@ -144,11 +144,11 @@ dydx.factor <- function(data, model, variable, type = c("response", "link"), fwr
     if (isTRUE(fwrap)) {
         outcolnames <- paste0("factor(", variable, ")", levs)
     } else {
-        outcolnames <- paste0(variable, levs)
+        outcolnames <- paste0("dydx_", variable, levs)
     }
     out <- structure(rep(list(list()), length(levs)), 
                      class = "data.frame", 
-                     names = paste0("dydx_",outcolnames), 
+                     names = outcolnames, 
                      row.names = seq_len(nrow(data)))
     
     # setup base data and prediction
@@ -192,7 +192,7 @@ dydx.logical <- function(data, model, variable, type = c("response", "link"), ..
     # calculate difference for moving FALSE to TRUE
     D1 <- build_datalist(data, at = setNames(list(TRUE), variable))[[1]]
     P1 <- prediction(model = model, data = D1, type = type)[["fitted"]]
-    out[[variable]] <- P1 - P0
+    out[[paste0("dydx_",variable)]] <- P1 - P0
     
     # return data.frame with column of differences
     class(out[[1]]) <- c("marginaleffect", "numeric")