+++ R/summary.R
@@ -1,12 +1,31 @@
 #' @export
 summary.margins <- 
-function(object, at = object[[".at"]], level = 0.95, ...) {
-    if (is.na(object[[".at"]][1])) {
-        summarize_one(object, level = level, ...)
+function(object, level = 0.95, order = "Factor", ...) {
+    if (is.null(attributes(object)[["at"]])) {
+        out <- summarize_one(object, level = level, ...)
+        out[[".at"]] <- NA_character_
+        out <- out[order(out[["Factor"]]),]
     } else {
-        tmp <- split(object, object[[".at"]])
-        lapply(tmp, summarize_one, level = level, ...)
+        object[[".at"]] <- unlist(lapply(object[[".at"]], collapse_at_vals, order = if (order == "Factor") NULL else order))
+        at_split <- split(object, object[[".at"]])
+        out <- list()
+        for (i in seq_along(at_split)) {
+            out[[i]] <- summarize_one(at_split[[i]], level = level, ...)
+            out[[i]][[".at"]] <- names(at_split)[i]
+        }
+        out <- do.call("rbind", out)
+        if (order[1L] == "Factor") {
+            out <- out[order(out[["Factor"]], out[[".at"]]),]
+        } else {
+            out <- out[order(out[[".at"]], out[["Factor"]]),]
+        }
     }
+    structure(out[ , c("Factor", ".at", names(out)[!names(out) %in% c("Factor", ".at")])], 
+              type = attributes(object)[["type"]],
+              call = attributes(object)[["call"]],
+              vce = attributes(object)[["vce"]],
+              iterations = attributes(object)[["iterations"]],
+              at = attributes(object)[["at"]])
 }
 
 summarize_one <- function(object, level = 0.95, ...) {
@@ -22,6 +41,5 @@ summarize_one <- function(object, level = 0.95, ...) {
     tab[["Pr(>|z|)"]] <- 2 * pnorm(abs(tab[,"z value"]), lower.tail = FALSE)
     structure(cbind(tab, confint(object = object, level = level)),
               class = c("summary.margins", "data.frame"),
-              call = attributes(object)[["call"]],
-              at = attributes(object)[["at"]])
+              call = attributes(object)[["call"]])
 }