+++ tests/testthat/tests-core.R
@@ -7,30 +7,30 @@ context("Basic accuracy tests")
 test_that("Test accuracy for lm()", {
     x <- lm(mpg ~ wt, data = mtcars)
     m <- marginal_effects(x)
-    expect_equal(coef(x)[["wt"]], mean(m[["wt"]]), tolerance = tol, label = "marginal effect is coefficient in lm()")
+    expect_equal(coef(x)[["wt"]], mean(m[["dydx_wt"]]), tolerance = tol, label = "marginal effect is coefficient in lm()")
 })
 test_that("Test accuracy for glm()", {
     x <- glm(am ~ wt, data = mtcars, family = binomial)
     m1 <- marginal_effects(x)
-    expect_true(coef(x)[["wt"]] != mean(m1[["wt"]]), label = "marginal effect (type = 'response') is not coefficient in glm()")
+    expect_true(coef(x)[["wt"]] != mean(m1[["dydx_wt"]]), label = "marginal effect (type = 'response') is not coefficient in glm()")
     m1b <- marginal_effects(x, type = "link")
-    expect_equal(coef(x)[["wt"]], mean(m1b[["wt"]]), tolerance = tol, label = "marginal effect (type = 'link') is coefficient in glm()")
+    expect_equal(coef(x)[["wt"]], mean(m1b[["dydx_wt"]]), tolerance = tol, label = "marginal effect (type = 'link') is coefficient in glm()")
     m2 <- marginal_effects(x, type = "link")
-    expect_equal(coef(x)[["wt"]], mean(m2[["wt"]]), tolerance = tol, label = "marginal effect is not coefficient in glm()")
+    expect_equal(coef(x)[["wt"]], mean(m2[["dydx_wt"]]), tolerance = tol, label = "marginal effect is not coefficient in glm()")
     p <- predict(x, type = "response")
     manual <- coef(x)[["wt"]] * p * (1-p)
-    expect_equal(as.numeric(manual), as.numeric(m1[["wt"]]), tolerance = tol, label = "marginal effect is correct for logit glm()")
+    expect_equal(as.numeric(manual), as.numeric(m1[["dydx_wt"]]), tolerance = tol, label = "marginal effect is correct for logit glm()")
 })
 test_that("Test accuracy for loess()", {
     x <- loess(mpg ~ wt, data = mtcars)
-    expect_true(inherits(m <- margins(x)[[1]], "margins"), label = "margins works for loess()")
+    expect_true(inherits(m <- margins(x), "margins"), label = "margins works for loess()")
 })
 
 
 context("Test `build_datalist()` behavior")
 test_that("Test build_datalist()", {
     expect_true(length(build_datalist(mtcars, at = list(cyl = c(4,6)))) == 2)
-    expect_true(length(build_datalist(mtcars, at = list(cyl = c(4,6), wt = c(1,1.5)))) == 4)
+    expect_true(length(build_datalist(mtcars, at = list(cyl = c(4,6), wt = c(1.6,1.8)))) == 4)
     m <- mtcars
     m[["cyl"]] <- factor(m[["cyl"]])
     expect_error(build_datalist(m, at = list(cyl = 10)), label = "factor error in build_datalist()")
@@ -80,14 +80,14 @@ context("print(), summary(), and confint() methods")
 test_that("print()/summary() for 'margins' object", {
     x <- lm(mpg ~ wt * hp, data = head(mtcars))
     m <- margins(x)
-    expect_true(inherits(print(m[[1]]), "margins"), label = "print() method for margins")
-    expect_true(inherits(summary(m[[1]]), "data.frame"), label = "summary() method for margins")
-    expect_true(inherits(print(summary(m[[1]])), "data.frame"), label = "print() method for summary.margins")
+    expect_true(inherits(print(m), "margins"), label = "print() method for margins")
+    expect_true(inherits(summary(m), "data.frame"), label = "summary() method for margins")
+    expect_true(inherits(print(summary(m)), "data.frame"), label = "print() method for summary.margins")
 })
 test_that("confint() for 'margins' object", {
     x <- lm(mpg ~ wt * hp, data = head(mtcars))
     m <- margins(x)
-    expect_true(inherits(confint(m[[1]]), "matrix"), label = "confint() for margins")
+    expect_true(inherits(confint(m), "matrix"), label = "confint() for margins")
 })
 
 
@@ -95,7 +95,10 @@ test_that("confint() for 'margins' object", {
 test_that("minimum test of variance calculations", {
     x <- lm(mpg ~ wt * hp, data = mtcars)
     expect_true(inherits(plot(margins(x, vce = "delta")), "margins"))
+    dev.off()
     expect_true(inherits(plot(margins(x, vce = "simulation", iter = 5L)), "margins"))
+    dev.off()
     expect_true(inherits(plot(margins(x, vce = "bootstrap", iter = 5L)), "margins"))
+    dev.off()
 })